+++ svd-parser/src/lib.rs
@@ -163,6 +163,7 @@ mod interrupt;
 mod modifiedwritevalues;
 mod peripheral;
 mod peripheralinfo;
+mod protection;
 mod readaction;
 mod register;
 mod registercluster;
@@ -204,6 +205,8 @@ pub enum SVDError {
     InvalidModifiedWriteValues(String),
     #[error("Invalid readAction variant, found {0}")]
     InvalidReadAction(String),
+    #[error("Invalid protection variant, found {0}")]
+    InvalidProtection(String),
     #[error("The content of the element could not be parsed to a boolean value {0}: {1}")]
     InvalidBooleanValue(String, core::str::ParseBoolError),
     #[error("dimIndex tag must contain {0} indexes, found {1}")]